Makefile.releases: Sort entries by version number
Sort entries in Makefile.releases by increasing version number.

The motivation is to be able to add new entries for patch releases from
multiple branches (vX.Y) without updating the same portion of the files,
and in particular, without creating conflicts for Pull Requests.

The file was updated with the Magic Command below:

    $ head -n 5 Makefile.releases > Makefile.releases.tmp \
    $ tail -n +5 Makefile.releases | tr '\n' '@' | \
        sed 's/@# Cilium/\n# Cilium/g' | sort -V | \
        sed -e '/^$/d' -e 's/@\+$/@@/' -e 's/@/\n/g' | \
        head --lines=-2 >> Makefile.releases.tmp
    $ mv Makefile.releases.tmp Makefile.releases

Signed-off-by: Quentin Monnet <quentin@isovalent.com>