add-release.sh: Preserve order for new entries in Makefile.releases
Now that we have sorted the entries in Makefile.releases by version
number, let's make sure that we preserve this order when inserting new
entries in the file.

The advantage over appending new entries to the document is that
handline patch releases for three stable branches will no longer update
the same portions of the file, and merge conflicts will be less likely
to occur if creating three Pull Requests in parallel.

Signed-off-by: Quentin Monnet <quentin@isovalent.com>