add-release.sh: Fix mapfile invocation
"mapfile" was added in response to shellcheck's complaints in CI, and
the change was not tested well enough. As it turns out, the current
invocation pulls the newline characters at the end of version numbers
into the array, making the version check fail.

We just miss the '-t' flag to automatically trim these newline
characters. Let's fix it.

Signed-off-by: Quentin Monnet <quentin@isovalent.com>