Add installation documentation for OKD
This commit ports installation instructions from the Cilium
documentation and adds it to a local file. This is part of an effort to
restructure OpenShift documentation within Cilium's documentation, after
the deprecation of cilium/cilium-olm. See cilium/cilium#24270.

The ported documentation includes a snippet containing information
on how to run the connectivity test suite. As this snippet is not
available in this repository, and to prevent having to keep two
duplicate sources of information up-to-date, the ported documentation
has been modified to point to Cilium's troubleshooting guide, which
contains instructions for running the connectivity tests.

Signed-off-by: Ryan Drew <ryan.drew@isovalent.com>